refactor: change folder structures for features folder
