+++ src/features/PokemonSearchPage/PokemonSearchPage.tsx
@@ -1,7 +1,7 @@
-import Search from '../../../components/Search/Search';
-import CardList from '../CardList/CardList';
-import ErrorBoundary from '../../../components/ErrorBoundary/ErrorBoundary';
-import useLocalStorage from '../../../hooks/useLocalStorage';
+import Search from '../../components/Search/Search';
+import CardList from '../PokemonList/CardList/CardList';
+import ErrorBoundary from '../../components/ErrorBoundary/ErrorBoundary';
+import useLocalStorage from '../../hooks/useLocalStorage';
 
 export default function PokemonSearchPage() {
   const [searchData, setSearchData] = useLocalStorage('searchInput', '');
